Add Refactored Dashboard and ReadingListView
Refactored the switcher, renaming it to Dashboard. The goal is to reduce
coupling between components in the Dashboard to make the code more
maintainable.

Reducing coupling meant controling the number of functions that need to
be passed through the various components in the views. Previously
functions were being created on the switcher object and passed down
through multiple levels of components until they reached their desired
users. This was hard to follow and maintain. The addition of redux
during this refactor made it so there was no coupling between the
components, the only requirement is that components that want to alter
the state connect to the redux store and define an action.

All of these changes required a rewrite of how the code is structured so
there are a lot of changes here. Since I was already re-writing the
components I decided to re-write their css as well. The goal was to make
it look as close to the original pre-refactor dashboard as possible.